makes Header folder in components/common and moves associated files
